Update celery to 5.2.2 and redis to 4.5.4.
These changes were suggested by depdendabot.

The celery update required a change to manager.py because the new celery
enforces a new order on parameters.

The .gitignore file is updated to track txt files and
`requirements.large.txt` was added to the repo. This file supports the
local Dockerfile testing strategy. These requirements take a long time
to build and do not change often, so they are placed in their own Docker
layer to shorten build times.